+++ views/register.hbs
@@ -2,7 +2,7 @@
     <form action="/register" method="POST">
         <div class="mb-3">
             <label for="email" class="form-label">Email address</label>
-            <input type="email" name= "email" class="form-control" id="email" aria-describedby="emailHelp">
+            <input type="email" value="{{user.email}}" name= "email" class="form-control {{#if errors.email}}is-invalid{{/if}}" id="email" aria-describedby="emailHelp" required>
             <div id="emailHelp" class="form-text">We'll never share your email with anyone else.</div>
             {{#if errors.email}}
                 <div class="invalid-feedback">
@@ -12,7 +12,7 @@
         </div>
         <div class="mb-3">
             <label for="exampleInputPassword1" class="form-label">Password</label>
-            <input type="password" name="password" class="form-control" id="exampleInputPassword1">
+            <input type="password" name="password" class="form-control {{#if errors.email}}is-invalid{{/if}}" id="exampleInputPassword1" required>
         </div>
             {{#if errors.password}}
                 <div class="invalid-feedback">
@@ -22,3 +22,5 @@
         <button type="submit" class="btn btn-primary">Submit</button>
     </form>
 </div>
+
+<a class= "btn" href="/">Home</a>
